whatsapp-bot/pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = "whatsapp bot"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
twilio = "^9.3.4"
fastapi = "^0.115.3"
python-dotenv = "^1.0.1"
python-multipart = "^0.0.12"
uvicorn = "^0.32.0"
pillow = "^11.0.0"
aiohttp = "^3.10.10"
google-cloud-vision = "^3.8.0"
aioredis = "^2.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
